Create Tharaka University Web Utitility
Accessing all web services from one point, and in a click of a button… The Tharaka University Application is a Web utility that is handy in providing all the services that are available in the University Website. It comes with a fully functional graphical user interface that features a web responsive, user friendly dash board with click buttons responding to all services available and a 24-hour format digital time that automatically matches that of the host device. It is easy to use and light weight, meaning that users don’t need any level of technical know how. It is a dynamic and flexible tool that can be modified to fit all institutions that offer web services. It is an exquisite tool that automates all the boring, tedious and redundant work of searching for URL links or websites. It is utility that you can never miss.

It is all what you need when it comes to web navigation, accessing all web services from one point, and in a click of a button.